requirements.txt
asgiref==3.4.1
astroid==2.8.4
auth==0.5.3
autopep8==1.6.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-multiselectfield==0.1.12
django-phonenumber-field==6.0.0
dnspython==2.1.0
eventlet==0.32.0
falcon==3.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mongoengine==0.23.1
phonenumberslite==8.12.36
platformdirs==2.4.0
pycodestyle==2.8.0
pylint==2.11.1
pymongo==3.12.1
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
